webapp/requirements.txt
Werkzeug==0.16.1
Flask==1.1.1
flask-restplus==0.13.0
psycopg2-binary
